fix(cache): share URL normalization between deduplicator and remove_url
Move url_to_hash/_normalize_url/_TRACKING_PARAMS from deduplicator.py
into cache.py as the single source of truth. cache.remove_url now uses
the same normalized hash the pipeline stores, so CLI cache remove works
correctly for URLs with tracking params or mixed case.

Also tightens test_cli_run_passes_config_to_pipeline to assert that the
config_path is correctly threaded into the Pipeline factory.
